requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
psycopg2-binary
django-environ